Move tests to the correct place.
