+++ pyxem/tests/test_signals/test_diffraction1d.py
@@ -18,6 +18,8 @@
 
 import numpy as np
 import dask.array as da
+import hyperspy.api as hs
+import pytest
 
 from pyxem.signals.diffraction1d import Diffraction1D, LazyDiffraction1D
 
@@ -78,3 +80,44 @@ class TestDecomposition:
         s = Diffraction1D(electron_diffraction1d)
         s.decomposition()
         assert isinstance(s, Diffraction1D)
+
+
+class TestVirtualImaging:
+    # Tests that virtual imaging runs without failure
+
+    @pytest.mark.parametrize('stack', [True, False])
+    def test_plot_interactive_virtual_image(self, stack,
+                                            electron_diffraction1d):
+        if stack:
+            electron_diffraction1d = hs.stack([electron_diffraction1d] * 3)
+        roi = hs.roi.SpanROI(left=1., right=2.)
+        electron_diffraction1d.plot_interactive_virtual_image(roi)
+
+    def test_get_virtual_image(self, electron_diffraction1d):
+        roi = hs.roi.SpanROI(left=1., right=2.)
+        vi = electron_diffraction1d.get_virtual_image(roi)
+        assert vi.data.shape == (2, 2)
+        assert vi.axes_manager.signal_dimension == 2
+        assert vi.axes_manager.navigation_dimension == 0
+
+    @pytest.mark.parametrize('out_signal_axes',
+                             [None, (0, 1), (1, 2), ('x', 'y')])
+    def test_get_virtual_image_stack(self, electron_diffraction1d,
+                                     out_signal_axes):
+        s = hs.stack([electron_diffraction1d] * 3)
+        s.axes_manager.navigation_axes[0].name = 'x'
+        s.axes_manager.navigation_axes[1].name = 'y'
+
+        roi = hs.roi.SpanROI(left=1., right=2.)
+        vi = s.get_virtual_image(roi, out_signal_axes)
+        assert vi.axes_manager.signal_dimension == 2
+        assert vi.axes_manager.navigation_dimension == 1
+        if out_signal_axes == (1, 2):
+            assert vi.data.shape == (2, 3, 2)
+            assert vi.axes_manager.navigation_size == 2
+            assert vi.axes_manager.signal_shape == (2, 3)
+        else:
+            assert vi.data.shape == (3, 2, 2)
+            assert vi.axes_manager.navigation_size == 3
+            assert vi.axes_manager.signal_shape == (2, 2)
+
