+++ pyxem/tests/test_signals/test_electron_diffraction1d.py
@@ -19,7 +19,6 @@
 import numpy as np
 import dask.array as da
 import pytest
-import hyperspy.api as hs
 
 from pyxem.signals.electron_diffraction1d import ElectronDiffraction1D
 from pyxem.signals.electron_diffraction1d import LazyElectronDiffraction1D
@@ -49,46 +48,6 @@ class TestSimpleHyperspy:
         electron_diffraction1d.set_diffraction_calibration(calibration)
         dx = electron_diffraction1d.axes_manager.signal_axes[0]
         assert dx.scale == calibration
-
-
-class TestVirtualImaging:
-    # Tests that virtual imaging runs without failure
-
-    @pytest.mark.parametrize('stack', [True, False])
-    def test_plot_interactive_virtual_image(self, stack,
-                                            electron_diffraction1d):
-        if stack:
-            electron_diffraction1d = hs.stack([electron_diffraction1d] * 3)
-        roi = hs.roi.SpanROI(left=1., right=2.)
-        electron_diffraction1d.plot_interactive_virtual_image(roi)
-
-    def test_get_virtual_image(self, electron_diffraction1d):
-        roi = hs.roi.SpanROI(left=1., right=2.)
-        vi = electron_diffraction1d.get_virtual_image(roi)
-        assert vi.data.shape == (2, 2)
-        assert vi.axes_manager.signal_dimension == 2
-        assert vi.axes_manager.navigation_dimension == 0
-
-    @pytest.mark.parametrize('out_signal_axes',
-                             [None, (0, 1), (1, 2), ('x', 'y')])
-    def test_get_virtual_image_stack(self, electron_diffraction1d,
-                                     out_signal_axes):
-        s = hs.stack([electron_diffraction1d] * 3)
-        s.axes_manager.navigation_axes[0].name = 'x'
-        s.axes_manager.navigation_axes[1].name = 'y'
-
-        roi = hs.roi.SpanROI(left=1., right=2.)
-        vi = s.get_virtual_image(roi, out_signal_axes)
-        assert vi.axes_manager.signal_dimension == 2
-        assert vi.axes_manager.navigation_dimension == 1
-        if out_signal_axes == (1, 2):
-            assert vi.data.shape == (2, 3, 2)
-            assert vi.axes_manager.navigation_size == 2
-            assert vi.axes_manager.signal_shape == (2, 3)
-        else:
-            assert vi.data.shape == (3, 2, 2)
-            assert vi.axes_manager.navigation_size == 3
-            assert vi.axes_manager.signal_shape == (2, 2)
 
 
 class TestComputeAndAsLazyElectron1D:
